* "Fixed" all warning messages, including signed/unsigned mismatch or unused variables. * Fixed: crash when closing the program with the main menu exit option. * Fixed: editing word in base dictionary didn't mark the dictionary user data as changed.

+++ import.h
@@ -169,7 +169,7 @@ public:
     bool error() const;
 
     // File size of the opened file.
-    int size() const;
+    qint64 size() const;
     // File position in the opened file.
     int pos() const;
     // Reads the next line into result and returns true if a line was read. Sets result to an
